applications/model_deployment_api/model-deployment-api-requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
flasgger==0.9.7b2
python-dotenv==1.0.1
scikit-learn==1.3.0
numpy==1.22.2
mlflow==2.14.2
folium==0.17.0
geopy==2.4.1
GeoAlchemy2==0.15.2
shapely==2.0.6
streamlit==1.40.0
streamlit-folium==0.22.1